fix(plugin-meetings): addressed review comments for moving cluster reachability functionality
